+++ Entities/Attachable.cpp
@@ -48,6 +48,7 @@ void Attachable::Clear()
 	m_InheritsRotAngle = true;
 	m_CanCollideWithTerrainWhenAttached = false;
 	m_IsCollidingWithTerrainWhileAttached = false;
+	m_DeleteWithParent = false;
 }
 
 
@@ -88,6 +89,7 @@ int Attachable::Create(const Attachable &reference)
     m_OnlyLinForces = reference.m_OnlyLinForces;
 	m_InheritsRotAngle = reference.m_InheritsRotAngle;
 	m_CanCollideWithTerrainWhenAttached = reference.m_CanCollideWithTerrainWhenAttached;
+	m_DeleteWithParent = reference.m_DeleteWithParent;
 
     return 0;
 }
@@ -119,6 +121,8 @@ int Attachable::ReadProperty(std::string propName, Reader &reader)
         reader >> m_DrawAfterParent;
     else if (propName == "CollidesWithTerrainWhenAttached")
         reader >> m_CanCollideWithTerrainWhenAttached;
+	else if (propName == "DeleteWithParent")
+		reader >> m_DeleteWithParent;
     else
         // See if the base class(es) can find a match instead
         return MOSRotating::ReadProperty(propName, reader);
@@ -153,6 +157,8 @@ int Attachable::Save(Writer &writer) const
     writer << m_DrawAfterParent;
     writer.NewProperty("CollidesWithTerrainWhenAttached");
     writer << m_CanCollideWithTerrainWhenAttached;
+	writer.NewProperty("DeleteWithParent");
+	writer << m_DeleteWithParent;
 
     return 0;
 }
@@ -242,17 +248,8 @@ bool Attachable::ParticlePenetration(HitData &hd)
 // Description:     Gibs this, effectively destroying it and creating multiple gibs or
 //                  pieces in its place.
 
-void Attachable::GibThis(Vector impactImpulse, float internalBlast, MovableObject *pIgnoreMO)
-{
-    if (m_pParent)
-    {
-        (MOSRotating *)m_pParent->RemoveAttachable(this);
-    }
-    else
-    {
-        Detach();
-    }
-
+void Attachable::GibThis(Vector impactImpulse, float internalBlast, MovableObject *pIgnoreMO) {
+    m_pParent ? m_pParent->RemoveAttachable(this) : Detach();
     MOSRotating::GibThis(impactImpulse, internalBlast, pIgnoreMO);
 }